# openapi.yaml
openapi: 3.0.0
info:
  description: Define an API
  title: Restful API
  version: 1.0.0
paths:
...
  "/{sql}/getalltables":
    get:
      tags:
        - Table Information
      summary: 取得所有資料表
      parameters:
        - description: 資料庫引擎
          name: sql
          in: path
          required: true
          schema:
            type: string
            example: 
              "mysql or mssql"
      responses:
        "200":
          description: Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.object"
              example: "列出所有資料表"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Error"
              example: 
                "message": "資料庫發生錯誤"
  "/{sql}/tableinformation/{tablename}":
    get:
      tags:
        - Table Information
      summary: 取得資料表資訊
      parameters:
        - description: 資料庫引擎
          name: sql
          in: path
          required: true
          schema:
            type: string
            example: 
              "mysql or mssql"
        - description: 資料表名稱
          name: tablename
          in: path
          required: true
          schema:
            type: string
          example: "peoples"
      responses:
        "200":
          description: Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.object"
              example:
                mysql:
                  Field:
                    "欄位名稱"
                  Type:
                    "欄位類型"                
                  Null:
                    "是否允許Null"
                  Key:
                    "key類型"
                  Default:
                    "預設值"
                  Extra: 
                    "額外資訊"
                mssql:
                  Field:
                    "欄位名稱"
                  Type:
                    "欄位類型"                
                  Null:
                    "是否允許Null"
                  Default:
                    "預設值"

        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Error"
              example:
                message: "資料庫發生錯誤"
...
  "/{sql}/getsome/{tablename}":
    get:
      tags:
        - Table CRUD
      summary: 取得部分資料
      parameters:
        - description: 資料庫引擎
          name: sql
          in: path
          required: true
          schema:
            type: string
            example: 
              "mysql or mssql"
        - description: 資料表名稱
          name: tablename
          in: path
          required: true
          schema:
            type: string
          example: peoples
        - description: 挑選欄位
          name: col
          in: query
          schema:
            type: array
            example: ["key1"]
            items:
              type: string
          style: form
          explode: true
        - description: 選擇條件(colname,value)
          name: where
          in: query
          schema:
            type: array
            example: ["key1,123456"]
            items:
              type: string
          style: form
          explode: true
      responses:
        "200":
          description: Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.object"
              example: 
                欄位名稱: "欄位數值"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Error"
              example:
                message: "資料庫發生錯誤"
...
  "/jointable/{uuid}":
    get:
      tags:
        - Join Table
      summary: 專屬uuid呼叫表
      parameters:
        - description: uuid
          name: uuid
          in: path
          required: true
          schema:
            type: string
          example: "bdf4bc5b-4d8b-414a-8adf-eaa02e987b7a"   
      responses:
        "200":
          description: Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.object"
              example: 
                欄位名稱: "欄位數值"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Error"
              example:
                message: "資料庫發生錯誤"
servers:
  - url: http://localhost:8080/v1
components:
  schemas:
...
    models.Error:
      type: object
      properties:
        message:
          type: string
    models.object:
      type: object